+++ changelog.md
@@ -11,6 +11,10 @@
 
 # Changelog
 
+version 3.3.0 - upcoming
+  * code folding for application/json
+  * fixed postman collection v2.1 import bug
+
 version 3.2.1
   * added option to not follow redirects
   * sticky response-tabs (dont reset to first tab anymore)